Refactor project structure and update dependencies
- Updated workspace members in Cargo.toml to include pain_bevy_visualizer instead of pain_graphics.
- Modified pain_app's Cargo.toml to replace pain_graphics dependency with pain_bevy_visualizer.
- Removed pain_graphics module and its associated files.
- Implemented a new visualization system in pain_bevy_visualizer using Bevy and Avian3D.
- Created a ParticlePlugin for managing simulation particles and their interactions.
- Added systems for updating particles and bonds, handling user input, and initializing the simulation.
- Updated main.rs in pain_bevy_visualizer to set up the application and run the simulation.
- Adjusted simulation parameters and resource management for better performance and clarity.

+++ pain_bevy_visualizer/src/main.rs
@@ -20,7 +20,7 @@ fn main() {
 }
 
 fn create_initial_simulation() -> SimulationState {
-    let mut sim_state = SimulationState::new(SIM_WIDTH, SIM_HEIGHT, SIM_DEPTH);
+    let mut sim_state = SimulationState::new(SIM_WIDTH as f32, SIM_HEIGHT as f32, SIM_DEPTH as f32);
     sim_state.initialize_classic_recipe();
     sim_state
 }
